+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - refactor(rgb-ring): add `RgbRingSequenceAction`, add `as_command` to all `RgbRingCommandAction`s
 - feat(speech-recognition): implement wake word detection, the wake word can be set with `UBO_WAKE_WORD` environment variable, the default is `hey pod` - closes #284
 - chore(rpc): remove protobuf files and generated code from the repository, they are now generated in ci/cd pipeline and added to the package and iso files.
+- refactor(speech-recognition): support multiple actions in an intent
 
 ## Version 1.3.0